Clean the working tree before testing first commit
This could be done by -c|--clean option, but this is virtually always
needed. Let's make it default.

As the current -c|--clean option becomes useless, the semantics is
changed. Specifying this would cause "git clean -xdff" every time before
testing a commit, not just before testing the first one in the range.